requirements.txt
gradio
cv2
numpy
matplotlib
deep-translator
transformers
pytorch
dotenv
logging
PIL
numpy